requirements/local.txt
-r base.txt # includes the base.txt requirements file

django-test-plus==1.0.7
